refaktor hjelpetekstVurdereMålgruppe.ts til tsx og endret gjenlevende.
